refactor: extract AnyJSON to shared Utilities, add .int case, fix agentId precision
